requirements.txt
python-dotenv
mysql-connector-python
psutil
py-cpuinfo
getmac
requests
speedtest-cli